haskell: RPATH and dynamic library management.
Since haskell packages tend to install their libraries in non-standard
directories (subdirs of ${libdir}), it would be tedious to configure
ld.so to find the right paths. Instead, rely on the consistency with
which ghc sets the RPATH with an absolute path into the HOST sysroot and
change it accordingly to match the target environment for binaries
shipped in each package.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>